fix: collaboration cursor chrome should not wrap mid-line (#1423)
Here be dragons!

The collaboration cursor caret needs to:
 - exist in the dom as a non-zero width element, so that when hovering over it, the label can display
 - yet, effectively not take up space in the dom, so that it doesn't cause wrapping or otherwise effect layout of the page

To achieve this, it took quite a bit of fiddling to figure out how to do this.

The caret is a span which has a before and after pseudo element.
The before element is what actually takes up space in the dom, and is colored via a border.
The color is actually set by reading the current color from the `.collaboration-cursor__caret` element. Allowing for dynamic coloring from JS.

There are a number of browser specific quirks with these hacks:
 - Firefox differs from Chrome & Safari in that it will split a word that is wrapping if not displayed as inline-block (whereas the others need display: inline)
 - Safari differs from Chrome & Firefox in that it needs the pseudo element to be position: relative to display a pseudo-element element with a negative margin

The word-joiner char (\u2060) is used to make sure the caret doesn't wrap around the text.

Therefore if modifying this code, please test in all major browsers to ensure that the caret is rendered correctly in all browsers.