feat: remove shiki dep, add new @blocknote/code-block package for slim shiki build (#1519)
This is a breaking change for syntax highlighting of code blocks, but removes any dependency on shiki languages. Now, to enable syntax highlighting for code blocks, you can use the new `codeBlock` option to customize your own shiki themes and languages with a highlighter you provide.

To make this as easy as possible, we've introduced an optional package `@blocknote/code-block` which will provide the most commonly used languages out of the box, in the most efficient way possible (dynamically loaded when the language is selected).