feat: re-implement Y.js collaboration as BlockNote plugins (#1638)
This re-implements `@tiptap/extension-collaboration` and `@tiptap/extension-collaboration-curosr` as BlockNote plugins.

This is also future-proofs us from the upcoming change to Tiptap's `@tiptap/y-tiptap` when we'd rather stick to `y-prosemirror`.

Along the way, I discovered that some of the patches being done in Tiptap are no longer required. So, this also was able to simplify things considerably