fix(react): destroy editor instances after two ticks
React's StrictMode makes it so that refs (which we use as the main trigger for mounting/unmounting the editor),
are triggered twice. This makes it difficult to know when the editor is _actually_ meant to be unmounted.
If an editor is unmounted during a render of React NodeViews, this can cause errors to occur (since elements which were in the dom at the start of the render are no longer in the dom by the time that the render occurs).

This change takes a similar approach to the editor integration in Tiptap, where it delays `unmount`ing the editor until 2 ticks have passed.
This is to allow React the opportunity to re-call `mount` with the same element, or if the ticks pass, to actually unmount the editor instance.

In most cases, delaying the removal of the editor instance should be fine, because it is not expected to be available after it has been destroyed.